+++ src/app/(protected)/[group]/[subGroup]/[instance]/(student)/my-preferences/page.tsx
