+++ src/app/(protected)/[group]/[subGroup]/[instance]/(student)/my-preferences/preference-selection.tsx
