+++ src/app/(protected)/[group]/[subGroup]/[instance]/(student)/student-overview.tsx
@@ -1,0 +1,7 @@
+export function StudentOverview({
+  params,
+}: {
+  params: { group: string; subGroup: string; instance: string };
+}) {
+  return <>add student details here</>;
+}
