+++ src/app/(protected)/[group]/[subGroup]/[instance]/(supervisor)/supervisor-overview.tsx
@@ -1,0 +1,7 @@
+export function SupervisorOverview({
+  params,
+}: {
+  params: { group: string; subGroup: string; instance: string };
+}) {
+  return <>add supervisor details here</>;
+}
